% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mid_error.R
\name{calc_mid_error}
\alias{calc_mid_error}
\title{calc_mid_error.}
\usage{
calc_mid_error(md = NULL, reconstructed_mid = NULL, best_r = NULL)
}
\arguments{
\item{md}{Normalized measured intensities}

\item{reconstructed_mid}{A reconstructed MID based on a true MID and a theoretical distribution.}

\item{best_r}{A named numeric vector of fragment ratios.}
}
\value{
A numeric vector of length(x).
}
\description{
\code{calc_mid_error} will compute the error of a theoretical and a estimated mid.
}
\examples{
\dontrun{
md <- c(50,100,10,0,50,50); md <- md/sum(md); names(md) <- -1:4
mid <- c(0.7,0,0,0,0.3); names(mid) <- paste0("M",0:4)
reconstructed_mid <- apply(t(CalcTheoreticalMDV(fml="C19H37NO4Si3")[1:5,1:6])*mid,1,sum)
best_r <- unlist(list("M+H"=0.9,"M+"=0.1))
CorMID:::calc_mid_error(md=md, reconstructed_mid=reconstructed_mid, best_r=best_r)
}
}
\keyword{internal}
