% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMID.R
\name{FitMID}
\alias{FitMID}
\title{FitMID.}
\usage{
FitMID(
  md = NULL,
  td = NULL,
  r = NULL,
  mid_fix = NULL,
  prec = 0.01,
  trace_steps = FALSE,
  penalize = NA
)
}
\arguments{
\item{md}{Measured distribution. Normalized (i.e. sum=1) raw intensity vector.}

\item{td}{Theoretical intensity distribution (using function 'CalcTheoreticalMDV').}

\item{r}{matrix of considere fragments and their potential occurence.}

\item{mid_fix}{May provide a numeric vector used as a given MID. Allows to estimate \code{r} individually.}

\item{prec}{Precision of the estimation of MID, set to 1\% as default.}

\item{trace_steps}{For testing purposes. Print the results of intermediate steps to console.}

\item{penalize}{Numeric exponent penalizing solutions with low M+H occurence. Formula is 1+3*(1-x)^penalty. NA to omit penalizing.}
}
\value{
Fitted MID with attributes.
}
\description{
\code{FitMID} will ...
}
\examples{
\dontrun{
int <- c(1560, 119203, 41927, 16932, 4438)
names(int) <- c(-2, 0, 1, 2, 3)
fml <- "C19H37NO4Si3"
td <- CalcTheoreticalMDV(fml=fml)
known_frags <- unlist(list("M+H"=0,"M+"=-1,"M-H"=-2,"M+H2O-CH4"=+2))
r <- matrix(rep(c(0,1), 4), nrow=2, dimnames=list(NULL, names(known_frags)))
FitMID(md=int, td=td, trace=TRUE, frag=c(-2, 0, 1, 2, 3), r=r)
}
}
\keyword{internal}
