% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isotopes}
\alias{isotopes}
\title{A data frame containing isotope information.}
\format{
A data frame of 5 columns for 308 chemical isotopes.
\describe{
\item{element}{The element name.}
\item{isotope}{The isotope name.}
\item{mass}{The absolute mass of this isotope in Dalton.}
\item{abundance}{The absolute abundance of this isotope.}
\item{ratioC}{The C ratio of this element.}
}
}
\source{
Imported from the enviPat package.
}
\usage{
isotopes
}
\description{
A data frame containing isotope information.
}
\keyword{datasets}
