% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{known_frags}
\alias{known_frags}
\title{The known fragments and their relative position to \code{[M+H]}.}
\format{
A named vector of the currently evaluated fragments.
}
\source{
Calculated using the internal CorMID function `get_idx_mat`.
}
\usage{
known_frags
}
\description{
The known fragments and their relative position to \code{[M+H]}.
}
\keyword{datasets}
