% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rawMID.R
\name{plot_rawMID}
\alias{plot_rawMID}
\title{plot_rawMID.}
\usage{
plot_rawMID(x = NULL, ...)
}
\arguments{
\item{x}{A raw mid vector.}

\item{...}{Used to pass arguments for plot to overwrite some default settings.}
}
\value{
NULL (plot is generated).
}
\description{
\code{plot_rawMID} will plot a raw mid with some color standards.
}
\examples{
fml <- "C9H20O3Si2"
attr(fml,"nmz") <- 5
attr(fml,"nbio") <- 3
# make up some fake measurement data for this based on assumed 10\% labelling at M3
mid <- c(0.9,0,0,0.1)
r <- unlist(list("M+H"=0.8, "M+"=0.1, "M+H2O-CH4"=0.1))
rMID <- CorMID:::ReconstructMID(mid=mid, r=r, fml=fml)
CorMID:::plot_rawMID(x = rMID)
CorMID:::plot_rawMID(x = rMID, ylim=c(0,max(rMID)))
CorMID:::plot_rawMID(x = rMID, xlim=c(-2,12), ylim=NULL, col=2, lwd=12, las=2, xlab="label")
}
\keyword{internal}
