% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precalc_idx}
\alias{precalc_idx}
\title{Pre-calculated index matrices to speed up calculations in function \code{getMID}.}
\format{
A list of 36 example metabolites.
}
\source{
Calculated using the internal CorMID function `get_idx_mat`.
}
\usage{
precalc_idx
}
\description{
Pre-calculated index matrices to speed up calculations in function \code{getMID}.
}
\keyword{datasets}
