% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructMID.R
\name{ReconstructMID}
\alias{ReconstructMID}
\title{ReconstructMID.}
\usage{
ReconstructMID(
  mid = NULL,
  r = unlist(list(`M+H` = 1)),
  fml = NULL,
  cutoff = 0.01
)
}
\arguments{
\item{mid}{A numeric vector with sum=1 and length of C atoms +1.}

\item{r}{Fragment ratios. A numeric vector with sum=1.}

\item{fml}{A compound formula.}

\item{cutoff}{Remove values below this threshold from output vector.}
}
\value{
A reconstructed MID.
}
\description{
\code{ReconstructMID} will reconstruct a measured GC-APCI-MS spectrum
 of a compound given its true MID and the fragment ratio.
}
\details{
tbd
}
\examples{
fml <- "C9H20O3Si2"
attr(fml,"nmz") <- 9
attr(fml,"nbio") <- 3
# make up some fake measurement data for this based on assumed 10\% labelling at M3
mid <- c(0.9,0,0,0.1)
r <- unlist(list("M+H"=0.8, "M+"=0.1, "M+H2O-CH4"=0.1))
(rMID <- CorMID:::ReconstructMID(mid=mid, r=r, fml=fml, cutoff=0.001))
CorMID:::plot_rawMID(rMID)
\dontrun{
CorMID::CorMID(int = rMID, fml=fml, prec=0.001, r=r, trace_steps = TRUE)
}
}
\keyword{internal}
