% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mid_error.R
\name{calc_mid_error}
\alias{calc_mid_error}
\title{calc_mid_error.}
\usage{
calc_mid_error(md = NULL, reconstructed_mid = NULL, best_r = NULL)
}
\arguments{
\item{md}{Normalized measured intensities}

\item{reconstructed_mid}{A reconstructed MID based on a true MID and a theoretical distribution.}

\item{best_r}{A named numeric vector of fragment ratios.}
}
\value{
A numeric vector of length(x).
}
\description{
\code{calc_mid_error} will compute the error of a theoretical and a estimated mid.
}
\keyword{internal}
