% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_errors.R
\name{weight_errors}
\alias{weight_errors}
\title{weight_errors.}
\usage{
weight_errors(rMpH = NULL, errs = NULL, penalize = NA, max_panel = 9)
}
\arguments{
\item{rMpH}{A numeric vector of fragment ratios for the M+H fragment.}

\item{errs}{A numeric vector of errors.}

\item{penalize}{A single numeric or NA to omit weighting.}

\item{max_panel}{Constant to modify weighting function}
}
\value{
A numeric vector of length(x).
}
\description{
\code{weight_errors} will apply a weighting function to a numeric vector of errors.
}
\keyword{internal}
